requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
sqlalchemy==2.0.15
python-multipart==0.0.6
aiofiles==23.1.0
pandas==2.0.2
pillow==9.5.0
requests==2.31.0
python-dotenv==1.0.0
